server/requirements.txt
Flask
Flask-Cors
requests
azure-storage-blob
python-dotenv
pandas
seaborn
matplotlib
statsmodels